refactor: move slack event from slack module to event module
