+++ bomplatform-slack/src/test/java/com/javabom/bomplatform/slack/event/message/SlackEventTest.java
@@ -1,5 +1,6 @@
 package com.javabom.bomplatform.slack.event.message;
 
+import com.javabom.bomplatform.event.message.SlackEvent;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
